Remove conflicting interop structs (#7561)
Removed our `RECT`, `HDC`, etc structs in favor of cswin32. This necessitated changing all of the existing P/Invokes back to `[DllImport]` as the `[LibaryImport]` doesn't work with the source generated types.

- Remove all `using foo =` workarounds.
- Push forward to cswin32 daily build to address bool cast issue
- New cswin32 uses System.Drawing types where plausible (`POINT` -> `Point`, as we currently do).

+++ src/System.Windows.Forms/src/System/Windows/Forms/DpiChangedEventArgs.cs
@@ -22,7 +22,7 @@ namespace System.Windows.Forms
             DeviceDpiOld = old;
             DeviceDpiNew = PARAM.SignedLOWORD(m.WParamInternal);
             Debug.Assert(PARAM.SignedHIWORD(m.WParamInternal) == DeviceDpiNew, "Non-square pixels!");
-            RECT suggestedRect = *(RECT*)m.LParamInternal;
+            RECT suggestedRect = *(RECT*)(nint)m.LParamInternal;
             SuggestedRectangle = Rectangle.FromLTRB(suggestedRect.left, suggestedRect.top, suggestedRect.right, suggestedRect.bottom);
         }
